backend/pyproject.toml
[tool.poetry]
name = "neuroresolv-backend"
version = "0.1.0"
description = "Adaptive AI Tutor & Accountability Partner for New Year Resolutions"
authors = ["Mahesh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
aiosqlite = "^0.22.1"
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.12.5"}
pydantic-settings = "^2.6.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.18"
google-adk = "^1.20.0"
opik = "^1.9.87"
chromadb = "^1.4.1"
pypdf = "^5.0.0"
ebooklib = "^0.18"
beautifulsoup4 = "^4.12.0"
httpx = "^0.28.0"
alembic = "^1.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
